+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Option to hide the module when nothing is playing, default: false (Issue #2)
 - Option to show the play count in the header, default: true
+- Option to update progress bars between data updates from the server
 
 ## [1.0.0] - 2019-10-14